src/2-model-service/requirements.txt
tensorflow==2.12.0
protobuf>=3.20.3,<4.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
psycopg2-binary>=2.9.3
pandas>=1.3.0